floats...

+++ src/main.c
@@ -37,13 +37,13 @@
 
 
 
-double polynomial_regression_train_and_test(double data[SAMPLE_SIZE][DATASET_FEATURES], int feature_size, int sample_size, int degree);
+float polynomial_regression_train_and_test(float data[SAMPLE_SIZE][DATASET_FEATURES], int feature_size, int sample_size, int degree);
 
 
 int main(void)
 {
 	uint32_t start_time, end_time;
-	double rmse;
+	float rmse;
 
 	DWT->CTRL |= DWT_CTRL_CYCCNTENA_Msk;
 	DWT->CYCCNT = 0;
@@ -66,14 +66,14 @@ int main(void)
 	while(1);
 }
 
-double polynomial_regression_train_and_test(double data[SAMPLE_SIZE][DATASET_FEATURES], int feature_size, int sample_size, int degree)
+float polynomial_regression_train_and_test(float data[SAMPLE_SIZE][DATASET_FEATURES], int feature_size, int sample_size, int degree)
 {
 	// Extract features and labels from input matrix
-	double **x = malloc(sample_size * sizeof(double *));
-	double *y = malloc(sample_size * sizeof(double));
+	float **x = malloc(sample_size * sizeof(float *));
+	float *y = malloc(sample_size * sizeof(float));
 	for (int i = 0; i < sample_size; i++)
 	{
-		x[i] = malloc((degree + 1) * sizeof(double));
+		x[i] = malloc((degree + 1) * sizeof(float));
 		for (int j = 0; j <= degree; j++)
 		{
 			x[i][j] = pow(data[i][0], j);
@@ -82,14 +82,14 @@ double polynomial_regression_train_and_test(double data[SAMPLE_SIZE][DATASET_FEA
 	}
 
 	// Compute x transpose matrix and x transpose times x matrix
-	double **x_transpose = malloc((degree + 1) * sizeof(double *));
-	double **x_transpose_x = malloc((degree + 1) * sizeof(double *));
-	double *x_transpose_y = malloc((degree + 1) * sizeof(double));
-	double *temp = malloc((degree + 1) * sizeof(double));
+	float **x_transpose = malloc((degree + 1) * sizeof(float *));
+	float **x_transpose_x = malloc((degree + 1) * sizeof(float *));
+	float *x_transpose_y = malloc((degree + 1) * sizeof(float));
+	float *temp = malloc((degree + 1) * sizeof(float));
 	for (int i = 0; i <= degree; i++)
 	{
-		x_transpose[i] = malloc(sample_size * sizeof(double));
-		x_transpose_x[i] = malloc((degree + 1) * sizeof(double));
+		x_transpose[i] = malloc(sample_size * sizeof(float));
+		x_transpose_x[i] = malloc((degree + 1) * sizeof(float));
 		for (int j = 0; j < sample_size; j++)
 		{
 			x_transpose[i][j] = x[j][i];
@@ -110,12 +110,12 @@ double polynomial_regression_train_and_test(double data[SAMPLE_SIZE][DATASET_FEA
 	}
 
 	// Solve the system of equations to get coefficients
-	double *coefficients = malloc((degree + 1) * sizeof(double));
+	float *coefficients = malloc((degree + 1) * sizeof(float));
 	for (int i = 0; i <= degree; i++)
 	{
 		for (int j = i + 1; j <= degree; j++)
 		{
-			double ratio = x_transpose_x[i][j] / x_transpose_x[i][i];
+			float ratio = x_transpose_x[i][j] / x_transpose_x[i][i];
 			for (int k = i; k <= degree; k++)
 			{
 				x_transpose_x[j][k] -= ratio * x_transpose_x[i][k];
@@ -126,7 +126,7 @@ double polynomial_regression_train_and_test(double data[SAMPLE_SIZE][DATASET_FEA
 
 	for (int i = degree; i >= 0; i--)
 	{
-		double sum = 0;
+		float sum = 0;
 		for (int j = i + 1; j <= degree; j++)
 		{
 			sum += x_transpose_x[i][j] * coefficients[j];
@@ -135,10 +135,10 @@ double polynomial_regression_train_and_test(double data[SAMPLE_SIZE][DATASET_FEA
 	}
 
 	// Calculate RMSE on training data
-	double rmse = 0;
+	float rmse = 0;
 	for (int i = 0; i < sample_size; i++)
 	{
-		double y_pred = 0;
+		float y_pred = 0;
 		for (int j = 0; j <= degree; j++)
 		{
 			y_pred += coefficients[j] * pow(data[i][0], j);
